Logic moved to BaseWorkflow, new unit test created
