+++ core/src/test/java/com/dtolabs/rundeck/core/execution/workflow/steps/node/impl/TestExecNodeStepExecutor.java
@@ -24,9 +24,9 @@
 package com.dtolabs.rundeck.core.execution.workflow.steps.node.impl;
 
 import com.dtolabs.rundeck.core.common.*;
-import com.dtolabs.rundeck.core.execution.*;
-import com.dtolabs.rundeck.core.execution.dispatch.Dispatchable;
-import com.dtolabs.rundeck.core.execution.dispatch.DispatcherResult;
+import com.dtolabs.rundeck.core.execution.ExecutionContext;
+import com.dtolabs.rundeck.core.execution.ExecutionContextImpl;
+import com.dtolabs.rundeck.core.execution.ExecutionException;
 import com.dtolabs.rundeck.core.execution.service.NodeExecutor;
 import com.dtolabs.rundeck.core.execution.service.NodeExecutorResult;
 import com.dtolabs.rundeck.core.execution.service.NodeExecutorResultImpl;
@@ -38,10 +38,7 @@ import com.dtolabs.rundeck.core.utils.FileUtils;
 
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
 
 /**
  * TestExecNodeStepExecutor is ...
@@ -192,243 +189,6 @@ public class TestExecNodeStepExecutor extends AbstractBaseTest {
             assertEquals(test1, testexec.testNode);
         }
 
-    }
-
-    public void testIgnoreErrorOnErrorHandlerWithKeepgoing() throws Exception {
-        final Framework frameworkInstance = getFrameworkInstance();
-        ExecNodeStepExecutor interpret = new ExecNodeStepExecutor(frameworkInstance);
-        ExecutionListener listener = new DummyExecutionListener();
-
-        //setup nodeexecutor for local node
-        testNodeExecutor testexec = new testNodeExecutor();
-        NodeExecutorService service = NodeExecutorService.getInstanceForFramework(getFrameworkInstance());
-        service.registerInstance("local", testexec);
-
-        //execute command interpreter on local node
-        final NodeEntryImpl test1 = new NodeEntryImpl("testhost", "test1");
-        final StepExecutionContext context = ExecutionContextImpl.builder()
-                .frameworkProject(PROJ_NAME)
-                .framework(frameworkInstance)
-                .user("blah")
-                .threadCount(1)
-                .executionListener(listener)
-                .build();
-        context.getExecutionListener();
-        final String[] strings = {"test", "command"};
-
-        //command with error handler
-        ExecCommand command = new ExecCommandBase() {
-            public String[] getCommand() {
-                return strings;
-            }
-            @Override
-            public StepExecutionItem getFailureHandler() {
-                return new HandlerExecutionItem() {
-                    @Override
-                    public boolean isKeepgoingOnSuccess() {
-                        return true;
-                    }
-                    @Override
-                    public String getType() {
-                        return null;
-                    }
-                    @Override
-                    public String getLabel() {
-                        return null;
-                    }
-                };
-            }
-        };
-        {
-            final NodeStepResult interpreterResult = interpret.executeNodeStep(context, command, test1);
-            assertTrue(((DummyExecutionListener)context.getExecutionListener()).isIgnoreErr());
-        }
-
-    }
-
-    public void testNotIgnoreErrorOnErrorHandlerWithKeepgoingFalse() throws Exception {
-        final Framework frameworkInstance = getFrameworkInstance();
-        ExecNodeStepExecutor interpret = new ExecNodeStepExecutor(frameworkInstance);
-        ExecutionListener listener = new DummyExecutionListener();
-
-        //setup nodeexecutor for local node
-        testNodeExecutor testexec = new testNodeExecutor();
-        NodeExecutorService service = NodeExecutorService.getInstanceForFramework(getFrameworkInstance());
-        service.registerInstance("local", testexec);
-
-        //execute command interpreter on local node
-        final NodeEntryImpl test1 = new NodeEntryImpl("testhost", "test1");
-        final StepExecutionContext context = ExecutionContextImpl.builder()
-                .frameworkProject(PROJ_NAME)
-                .framework(frameworkInstance)
-                .user("blah")
-                .threadCount(1)
-                .executionListener(listener)
-                .build();
-        context.getExecutionListener();
-        final String[] strings = {"test", "command"};
-
-        //command with error handler
-        ExecCommand command = new ExecCommandBase() {
-            public String[] getCommand() {
-                return strings;
-            }
-            @Override
-            public StepExecutionItem getFailureHandler() {
-                return new HandlerExecutionItem() {
-                    @Override
-                    public boolean isKeepgoingOnSuccess() {
-                        return false;
-                    }
-                    @Override
-                    public String getType() {
-                        return null;
-                    }
-                    @Override
-                    public String getLabel() {
-                        return null;
-                    }
-                };
-            }
-        };
-        {
-            final NodeStepResult interpreterResult = interpret.executeNodeStep(context, command, test1);
-            assertFalse(((DummyExecutionListener)context.getExecutionListener()).isIgnoreErr());
-        }
-
-    }
-
-    public void testNotIgnoreErrorWithoutErrorHandler() throws Exception {
-        final Framework frameworkInstance = getFrameworkInstance();
-        ExecNodeStepExecutor interpret = new ExecNodeStepExecutor(frameworkInstance);
-        ExecutionListener listener = new DummyExecutionListener();
-
-        //setup nodeexecutor for local node
-        testNodeExecutor testexec = new testNodeExecutor();
-        NodeExecutorService service = NodeExecutorService.getInstanceForFramework(getFrameworkInstance());
-        service.registerInstance("local", testexec);
-
-        //execute command interpreter on local node
-        final NodeEntryImpl test1 = new NodeEntryImpl("testhost", "test1");
-        final StepExecutionContext context = ExecutionContextImpl.builder()
-                .frameworkProject(PROJ_NAME)
-                .framework(frameworkInstance)
-                .user("blah")
-                .threadCount(1)
-                .executionListener(listener)
-                .build();
-        context.getExecutionListener();
-        final String[] strings = {"test", "command"};
-
-        //command with error handler
-        ExecCommand command = new ExecCommandBase() {
-            public String[] getCommand() {
-                return strings;
-            }
-        };
-        {
-            final NodeStepResult interpreterResult = interpret.executeNodeStep(context, command, test1);
-            assertFalse(((DummyExecutionListener)context.getExecutionListener()).isIgnoreErr());
-        }
-
-    }
-
-    class DummyExecutionListener implements ExecutionListener {
-        boolean ignoreErr = false;
-        @Override
-        public void ignoreErrors(boolean ignore) {
-            ignoreErr=ignore;
-        }
-        @Override
-        public boolean isTerse() {
-            return false;
-        }
-
-        @Override
-        public String getLogFormat() {
-            return null;
-        }
-
-        @Override
-        public void log(int level, String message) {
-
-        }
-
-        @Override
-        public void event(String eventType, String message, Map eventMeta) {
-
-        }
-
-        @Override
-        public FailedNodesListener getFailedNodesListener() {
-            return null;
-        }
-
-        @Override
-        public void beginNodeExecution(ExecutionContext context, String[] command, INodeEntry node) {
-
-        }
-
-        @Override
-        public void finishNodeExecution(NodeExecutorResult result, ExecutionContext context, String[] command, INodeEntry node) {
-
-        }
-
-        @Override
-        public void beginNodeDispatch(ExecutionContext context, StepExecutionItem item) {
-
-        }
-
-        @Override
-        public void beginNodeDispatch(ExecutionContext context, Dispatchable item) {
-
-        }
-
-        @Override
-        public void finishNodeDispatch(DispatcherResult result, ExecutionContext context, StepExecutionItem item) {
-
-        }
-
-        @Override
-        public void finishNodeDispatch(DispatcherResult result, ExecutionContext context, Dispatchable item) {
-
-        }
-
-        @Override
-        public void beginFileCopyFileStream(ExecutionContext context, InputStream input, INodeEntry node) {
-
-        }
-
-        @Override
-        public void beginFileCopyFile(ExecutionContext context, File input, INodeEntry node) {
-
-        }
-
-        @Override
-        public void beginFileCopyFile(ExecutionContext context, List<File> files, INodeEntry node) {
-
-        }
-
-        @Override
-        public void beginFileCopyScriptContent(ExecutionContext context, String input, INodeEntry node) {
-
-        }
-
-        @Override
-        public void finishFileCopy(String result, ExecutionContext context, INodeEntry node) {
-
-        }
-
-        @Override
-        public void finishMultiFileCopy(String[] result, ExecutionContext context, INodeEntry node) {
-
-        }
-
-        @Override
-        public ExecutionListenerOverride createOverride() {
-            return null;
-        }
-        public boolean isIgnoreErr(){return ignoreErr;}
     }
 
 }
