+++ core/src/test/java/com/dtolabs/rundeck/core/execution/workflow/steps/node/impl/TestScriptURLNodeStepExecutor.java
@@ -26,9 +26,9 @@ package com.dtolabs.rundeck.core.execution.workflow.steps.node.impl;
 import com.dtolabs.rundeck.core.common.*;
 import com.dtolabs.rundeck.core.common.impl.URLFileUpdater;
 import com.dtolabs.rundeck.core.dispatcher.DataContextUtils;
-import com.dtolabs.rundeck.core.execution.*;
-import com.dtolabs.rundeck.core.execution.dispatch.Dispatchable;
-import com.dtolabs.rundeck.core.execution.dispatch.DispatcherResult;
+import com.dtolabs.rundeck.core.execution.ExecutionContext;
+import com.dtolabs.rundeck.core.execution.ExecutionContextImpl;
+import com.dtolabs.rundeck.core.execution.StepExecutionItem;
 import com.dtolabs.rundeck.core.execution.service.*;
 import com.dtolabs.rundeck.core.execution.workflow.StepExecutionContext;
 import com.dtolabs.rundeck.core.execution.workflow.steps.FailureReason;
@@ -604,376 +604,5 @@ public class TestScriptURLNodeStepExecutor extends AbstractBaseTest {
                 stringMapMap);
             assertEquals("http://example.com/path/node%2Fname?query=some%20value%20%3F%20for%20things%20%26%20stuff", value);
         }
-    }
-
-
-    public void testIgnoreErrorOnErrorHandlerWithKeepgoing() throws Exception {
-        final Framework frameworkInstance = getFrameworkInstance();
-        ScriptURLNodeStepExecutor interpret = new ScriptURLNodeStepExecutor(frameworkInstance);
-        ExecutionListener listener = new DummyExecutionListener();
-
-        //setup nodeexecutor for local node
-        multiTestNodeExecutor testexec = new multiTestNodeExecutor();
-        NodeExecutorService service = NodeExecutorService.getInstanceForFramework(getFrameworkInstance());
-        service.registerInstance("local", testexec);
-
-        testFileCopier testcopier = new testFileCopier();
-        FileCopierService copyservice = FileCopierService.getInstanceForFramework(getFrameworkInstance());
-        copyservice.registerInstance("local", testcopier);
-
-        //execute command interpreter on local node
-        final NodeEntryImpl test1 = new NodeEntryImpl("testhost1", "test1");
-        test1.setOsFamily("unix");
-
-        final StepExecutionContext context = ExecutionContextImpl.builder()
-                .frameworkProject(PROJ_NAME)
-                .framework(frameworkInstance)
-                .user("blah")
-                .executionListener(listener)
-                .build();
-        final String urlString = "http://test.com";
-
-        ScriptURLCommandBase command = new ScriptURLCommandBase() {
-            public String getURLString() {
-                return urlString;
-            }
-
-            public String[] getArgs() {
-                return new String[0];
-            }
-
-            @Override
-            public StepExecutionItem getFailureHandler() {
-                return new HandlerExecutionItem() {
-                    @Override
-                    public boolean isKeepgoingOnSuccess() {
-                        return true;
-                    }
-                    @Override
-                    public String getType() {
-                        return null;
-                    }
-                    @Override
-                    public String getLabel() {
-                        return null;
-                    }
-                };
-            }
-
-            public boolean isKeepgoingOnSuccess() {
-                return false;
-            }
-
-            public String getScriptInterpreter() {
-                return null;
-            }
-
-            @Override
-            public String getFileExtension() {
-                return null;
-            }
-
-            public boolean getInterpreterArgsQuoted() {
-                return false;
-            }
-
-        };
-        {
-            final ArrayList<NodeExecutorResult> nodeExecutorResults = new ArrayList<NodeExecutorResult>();
-            nodeExecutorResults.add(NodeExecutorResultImpl.createSuccess(null));
-            nodeExecutorResults.add(NodeExecutorResultImpl.createSuccess(null));
-            nodeExecutorResults.add(NodeExecutorResultImpl.createSuccess(null));
-            testexec.testResult = nodeExecutorResults;
-            testcopier.testResult = "/test/file/path";
-            final test1 interaction = new TestScriptURLNodeStepExecutor.test1();
-
-            interaction.httpResultCode = 200;
-            interaction.httpStatusText = "OK";
-            interaction.responseHeaders.put("Content-Type", new Header("Content-Type", "text/plain"));
-            String testcontent = "test script content";
-            ByteArrayInputStream stringStream = new ByteArrayInputStream(testcontent.getBytes());
-            interaction.bodyStream = stringStream;
-
-            interpret.setInteraction(interaction);
-
-            final NodeStepResult interpreterResult = interpret.executeNodeStep(context, command, test1);
-
-            assertTrue(((DummyExecutionListener)context.getExecutionListener()).isIgnoreErr());
-        }
-    }
-
-    public void testNoIgnoreErrorOnErrorHandlerWithKeepgoingFalse() throws Exception {
-        final Framework frameworkInstance = getFrameworkInstance();
-        ScriptURLNodeStepExecutor interpret = new ScriptURLNodeStepExecutor(frameworkInstance);
-        ExecutionListener listener = new DummyExecutionListener();
-
-        //setup nodeexecutor for local node
-        multiTestNodeExecutor testexec = new multiTestNodeExecutor();
-        NodeExecutorService service = NodeExecutorService.getInstanceForFramework(getFrameworkInstance());
-        service.registerInstance("local", testexec);
-
-        testFileCopier testcopier = new testFileCopier();
-        FileCopierService copyservice = FileCopierService.getInstanceForFramework(getFrameworkInstance());
-        copyservice.registerInstance("local", testcopier);
-
-        //execute command interpreter on local node
-        final NodeEntryImpl test1 = new NodeEntryImpl("testhost1", "test1");
-        test1.setOsFamily("unix");
-
-        final StepExecutionContext context = ExecutionContextImpl.builder()
-                .frameworkProject(PROJ_NAME)
-                .framework(frameworkInstance)
-                .user("blah")
-                .executionListener(listener)
-                .build();
-        final String urlString = "http://test.com";
-
-        ScriptURLCommandBase command = new ScriptURLCommandBase() {
-            public String getURLString() {
-                return urlString;
-            }
-
-            public String[] getArgs() {
-                return new String[0];
-            }
-
-            @Override
-            public StepExecutionItem getFailureHandler() {
-                return new HandlerExecutionItem() {
-                    @Override
-                    public boolean isKeepgoingOnSuccess() {
-                        return false;
-                    }
-                    @Override
-                    public String getType() {
-                        return null;
-                    }
-                    @Override
-                    public String getLabel() {
-                        return null;
-                    }
-                };
-            }
-
-            public boolean isKeepgoingOnSuccess() {
-                return false;
-            }
-
-            public String getScriptInterpreter() {
-                return null;
-            }
-
-            @Override
-            public String getFileExtension() {
-                return null;
-            }
-
-            public boolean getInterpreterArgsQuoted() {
-                return false;
-            }
-
-        };
-        {
-            final ArrayList<NodeExecutorResult> nodeExecutorResults = new ArrayList<NodeExecutorResult>();
-            nodeExecutorResults.add(NodeExecutorResultImpl.createSuccess(null));
-            nodeExecutorResults.add(NodeExecutorResultImpl.createSuccess(null));
-            nodeExecutorResults.add(NodeExecutorResultImpl.createSuccess(null));
-            testexec.testResult = nodeExecutorResults;
-            testcopier.testResult = "/test/file/path";
-            final test1 interaction = new TestScriptURLNodeStepExecutor.test1();
-
-            interaction.httpResultCode = 200;
-            interaction.httpStatusText = "OK";
-            interaction.responseHeaders.put("Content-Type", new Header("Content-Type", "text/plain"));
-            String testcontent = "test script content";
-            ByteArrayInputStream stringStream = new ByteArrayInputStream(testcontent.getBytes());
-            interaction.bodyStream = stringStream;
-
-            interpret.setInteraction(interaction);
-
-            final NodeStepResult interpreterResult = interpret.executeNodeStep(context, command, test1);
-
-            assertFalse(((DummyExecutionListener)context.getExecutionListener()).isIgnoreErr());
-        }
-    }
-
-    public void testNoIgnoreErrorWithoutErrorHandler() throws Exception {
-        final Framework frameworkInstance = getFrameworkInstance();
-        ScriptURLNodeStepExecutor interpret = new ScriptURLNodeStepExecutor(frameworkInstance);
-        ExecutionListener listener = new DummyExecutionListener();
-
-        //setup nodeexecutor for local node
-        multiTestNodeExecutor testexec = new multiTestNodeExecutor();
-        NodeExecutorService service = NodeExecutorService.getInstanceForFramework(getFrameworkInstance());
-        service.registerInstance("local", testexec);
-
-        testFileCopier testcopier = new testFileCopier();
-        FileCopierService copyservice = FileCopierService.getInstanceForFramework(getFrameworkInstance());
-        copyservice.registerInstance("local", testcopier);
-
-        //execute command interpreter on local node
-        final NodeEntryImpl test1 = new NodeEntryImpl("testhost1", "test1");
-        test1.setOsFamily("unix");
-
-        final StepExecutionContext context = ExecutionContextImpl.builder()
-                .frameworkProject(PROJ_NAME)
-                .framework(frameworkInstance)
-                .user("blah")
-                .executionListener(listener)
-                .build();
-        final String urlString = "http://test.com";
-
-        ScriptURLCommandBase command = new ScriptURLCommandBase() {
-            public String getURLString() {
-                return urlString;
-            }
-
-            public String[] getArgs() {
-                return new String[0];
-            }
-
-            @Override
-            public StepExecutionItem getFailureHandler() {
-                return null;
-            }
-
-            public boolean isKeepgoingOnSuccess() {
-                return false;
-            }
-
-            public String getScriptInterpreter() {
-                return null;
-            }
-
-            @Override
-            public String getFileExtension() {
-                return null;
-            }
-
-            public boolean getInterpreterArgsQuoted() {
-                return false;
-            }
-
-        };
-        {
-            final ArrayList<NodeExecutorResult> nodeExecutorResults = new ArrayList<NodeExecutorResult>();
-            nodeExecutorResults.add(NodeExecutorResultImpl.createSuccess(null));
-            nodeExecutorResults.add(NodeExecutorResultImpl.createSuccess(null));
-            nodeExecutorResults.add(NodeExecutorResultImpl.createSuccess(null));
-            testexec.testResult = nodeExecutorResults;
-            testcopier.testResult = "/test/file/path";
-            final test1 interaction = new TestScriptURLNodeStepExecutor.test1();
-
-            interaction.httpResultCode = 200;
-            interaction.httpStatusText = "OK";
-            interaction.responseHeaders.put("Content-Type", new Header("Content-Type", "text/plain"));
-            String testcontent = "test script content";
-            ByteArrayInputStream stringStream = new ByteArrayInputStream(testcontent.getBytes());
-            interaction.bodyStream = stringStream;
-
-            interpret.setInteraction(interaction);
-
-            final NodeStepResult interpreterResult = interpret.executeNodeStep(context, command, test1);
-
-            assertFalse(((DummyExecutionListener)context.getExecutionListener()).isIgnoreErr());
-        }
-    }
-
-    class DummyExecutionListener implements ExecutionListener {
-        boolean ignoreErr = false;
-        @Override
-        public void ignoreErrors(boolean ignore) {
-            ignoreErr=ignore;
-        }
-        @Override
-        public boolean isTerse() {
-            return false;
-        }
-
-        @Override
-        public String getLogFormat() {
-            return null;
-        }
-
-        @Override
-        public void log(int level, String message) {
-
-        }
-
-        @Override
-        public void event(String eventType, String message, Map eventMeta) {
-
-        }
-
-        @Override
-        public FailedNodesListener getFailedNodesListener() {
-            return null;
-        }
-
-        @Override
-        public void beginNodeExecution(ExecutionContext context, String[] command, INodeEntry node) {
-
-        }
-
-        @Override
-        public void finishNodeExecution(NodeExecutorResult result, ExecutionContext context, String[] command, INodeEntry node) {
-
-        }
-
-        @Override
-        public void beginNodeDispatch(ExecutionContext context, StepExecutionItem item) {
-
-        }
-
-        @Override
-        public void beginNodeDispatch(ExecutionContext context, Dispatchable item) {
-
-        }
-
-        @Override
-        public void finishNodeDispatch(DispatcherResult result, ExecutionContext context, StepExecutionItem item) {
-
-        }
-
-        @Override
-        public void finishNodeDispatch(DispatcherResult result, ExecutionContext context, Dispatchable item) {
-
-        }
-
-        @Override
-        public void beginFileCopyFileStream(ExecutionContext context, InputStream input, INodeEntry node) {
-
-        }
-
-        @Override
-        public void beginFileCopyFile(ExecutionContext context, File input, INodeEntry node) {
-
-        }
-
-        @Override
-        public void beginFileCopyFile(ExecutionContext context, List<File> files, INodeEntry node) {
-
-        }
-
-        @Override
-        public void beginFileCopyScriptContent(ExecutionContext context, String input, INodeEntry node) {
-
-        }
-
-        @Override
-        public void finishFileCopy(String result, ExecutionContext context, INodeEntry node) {
-
-        }
-
-        @Override
-        public void finishMultiFileCopy(String[] result, ExecutionContext context, INodeEntry node) {
-
-        }
-
-        @Override
-        public ExecutionListenerOverride createOverride() {
-            return null;
-        }
-        public boolean isIgnoreErr(){return ignoreErr;}
     }
 }
